requirements.txt
langgraph>=0.2.30
google-genai>=0.3.0
yfinance>=0.2.54
pandas>=2.2.2
lxml>=5.2.2
python-dotenv>=1.0.1
